Improve wordwrapping for code blocks in markdown
This improves one line code blocks in that it no longer reformats to put
the code on the same line as the ``` which indicates the start of the
block when you run 'gq'.

However it still wraps multiline code blocks.

This is set in an after config file because otherwise it is overwritten
by vim's built in markdown plugin